Produce a nice listing of messages once connected
e.g

+OK Mailbox open, 4 messages
   1: Test
       Andrew Clayton <andrew@digital-domain.net>
       Sun, 28 Apr 2013 12:20:09 +0100
   2: Test
       Andrew Clayton <andrew@digital-domain.net>
       Sun, 28 Apr 2013 00:28:46 +0100
   3: Test
       Andrew Clayton <andrew@digital-domain.net>
       Sun, 28 Apr 2013 00:04:55 +0100

When generating the list we use the TOP command to get the message
headers. We don't know how long the headers are so don't know how much
we need to read from the socket.

Currently I'm just looping for (;;) { reading from the socket until we
read in "\r\n\r\n.\r\n" } which would signify the end of the message
headers (at least in the POP server I'm testing against, my ISP's).

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>