Re-create the message list after a DELE command
When a message is deleted, it is only marked for deletion (RSET will
bring it back). So you send up with something like

    1 1118
    3 1100

after the 2nd message is marked for deletion. This caused a problem in
get_message_hdrs() as we used message - 1 for the message headers
structure array index. As in this case when it came to the second
message it would be trying to access msg_hdrs[2], which is invalid
memory.

So we need to use nr_messages - 1 as the array index, as this value
isn't linked to the message number.

We also need to reinitialise the msg_hdrs strcuture and nr_messages
counter in free_msg_hdrs().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>