Some code optimisation in get_message_hdrs()
In get_message_hdrs() we read from the network socket until we see a
"\r\n\r\n.\r\n" sequence which should denote the end of the headers.

As part of this loop we do, rlen += bytes_read, if we do that last then
we can search for the above sequence in buf + rlen rather than the whole
of buf, i.e just search the bit we just read.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>