Use read_pop_response_sync() in get_message_list()
We need to modify the string we look for to know when to stop reading in
read_pop_response_sync() a little.

Previously we were looking for "\r\n\r\n.\r\n" which is fine for the
mail headers but for the message list, we get a response like

    1 1118
    2 1025
    3 1878
    4 1116
    5 39900
    6 1100
    .

If we relax the string slightly to "\r\n.\r\n", this catches both cases.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>