Supply both arguments to the TOP command
The TOP command is used to display the headers and optionally some lines
from the body of the specified message. Usage is like; top n [l], where n
is the message number and l is the number of lines of the body to show.

We were just sending TOP n in get_message_hdrs()

It seems however that some POP servers, e.g The University of
Washington's, require the second argument to also be given.

So we now send; top n 0

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>