Re-work read_pop_response_sync()
evilxyz on GitHub reported an issue[0] whereby upon logging into the pop
server, popx would start consuming 100% cpu.

This was happening during the retrieval of the message list when the
list data exceeded the buffer size (8192). We would then end up in a
loop reading 0 bytes.

This was due to the count argument to read(2) being calculated as
count - total which would get down to zero but we still hadn't seen the
CRLF.CRLF terminator yet.

Re-work this function so that we now allocate the buffer (as needed) in
this function. We don't pass in any count argument and just read in
BUF_SIZE-1 (for the terminating nul byte) in each read() iteration
reallocating the buffer each time round as needed.

This will allow us to read in any amount of data that is required.

Also remove the fd argument as that wasn't even being used. Finally
rename the function to be a bit more meaningful, so is now looks like

    ssize_t read_pop_response_multi_line(char **buf)

This also allows us to remove the round_bytes_up() as that's no longer
used.

[0] - https://github.com/ac000/popx/issues/1

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>