Display message header downloading progress
Depending on the POP server, network conditions and number of messages,
it can take a while to download the headers and before anything is
displayed back to the user after the initial login.

In get_message_list() display a progress line saying what message
numbers headers is currently being downloaded.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>