Store the message headers in a binary tree
Rather than store the message header structs (msg_hdr) in an array,
store them in a binary tree.

This makes use of the glibc TSEARCH(3) API.

This will allow for some optimisations. The first of which is a change
to how the DELE command works. This change was also needed to avoid a
segmentation fault where we would try to retrieve and display an entry
from the binary tree that no longer existed.

When a message is deleted it is only marked for deletion and the pop
server will just skip showing it, messages are not renumbered. So if you
had 5 messages; 1, 2, 3, 4, & 5 and deleted message 3, you would then
only see; 1, 2, 4, & 5.

So now rather then regenerating the message list when ever a delete
happens, we mimic the pop server and just mark in the messages struct
msg_hdr that is has been marked for deletion and skip showing it.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>