Properly handle the RSET command
The RSET command is used to restore the POP session back to its initial
state. e.g if any messages had been marked for deletion they would be
unmarked.

Previously we just passed the command straight through to the server,
now we also handle it in popx and unmark any messages marked for
deletion in the binary tree. This keeps popx and the server in sync.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>